requirements.txt
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
exceptiongroup==1.2.2
idna==3.10
importlib-metadata==8.5.0
iniconfig==2.0.0
multimethod==1.10
mypy-extensions==1.0.0
numpy==1.24.4
packaging==24.1
pandas==2.0.3
pandera==0.20.4
pluggy==1.5.0
pydantic==2.9.2
pydantic-core==2.23.4
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
tomli==2.0.2
typeguard==4.3.0
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
urllib3==2.2.3
wrapt==1.16.0
zipp==3.20.2
